Удалил лишний уровень наследования. + Мелкие изменения в генерации маппера.
